+++ src/audio/ISoundEffect.hpp
@@ -7,10 +7,10 @@
 
 class ISoundEffect : public IRefCounted {
 public:
-    virtual void  play() = 0;
-    virtual void  stop() = 0;
-    virtual float getVolume() const = 0;
-    virtual void  setVolume(float volume) = 0;
+    virtual void play() = 0;
+    virtual void stop() = 0;
+    virtual int  getVolume() const = 0;
+    virtual void setVolume(int volume) = 0;
 
 protected:
     ~ISoundEffect() { }
